move much of peers component to peersService
